Get rid of compiler warning. JIRA: CXX-563
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@38822 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objtools/data_loaders/genbank/processors.cpp
@@ -1220,7 +1220,7 @@ void CProcessor_St_SE_SNPT::ProcessStream(CReaderRequestResult& result,
         LogStat(result, r, blob_id,
                 CGBRequestStatistics::eStat_LoadSNPBlob,
                 "CProcessor_St_SE_SNPT: read SNP table",
-                size);
+                double(size));
     }}
     
     CWriter* writer = GetWriter(result);
